Lots of stuff
I did a ton of stuff, so I'm just going to upload it all. I also updated the project to 2019.4, so there's that. I've made UI elements in the Main Menu Scene, and am working on getting idle anims for all characters, making a minimap, and giving the Enemys brains